refactored downbeat and pattern stuff into features.downbeats
